Initial commit for OVC2
OVC2 is a rewrite of the OVC (Open Video Chat) project. The OVC project
is aimed at running on the XO laptops.

The aims of this project:

 	- Run on all XOs.
	- Use entirely open source software.

The build system uses gnu autotools (argg painful, I know). Yes, its a pain
to get working and what not, but I am hoping its worth it in the long run.

Signed-off-by: Alex Waterman <imNotListening@gmail.com>